Fixes for you, my master. Додав папку models, файл db.py в цій папці, доналаштував компоуз. Добавив рядок "EXPOSE 8000" в докерфайл, ніякого результату. Намагався слідувати твоїй рекомендації щодо запуска сервера через докер, нічого не вийшло, тому забиваємо хуй.
