+++ .dockerignore
@@ -1,4 +1,3 @@
 venv
 .env.sample
 .idea
-.env
