+++ docker-compose.yml
@@ -1,11 +1,18 @@
 version: '3.11'
 services:
+  app:
+    image: app
+    ports:
+      - "8000:8000"
+    build:
+      context: .
+
   db:
     image: postgres
     restart: always
     environment:
-      - POSTGRES_USER=postgres
-      - POSTGRES_PASSWORD=postgres
+      - POSTGRES_USER=postgresql
+      - POSTGRES_PASSWORD=postgresql
     ports:
       - '5432:5432'
     volumes:
